# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PlanIt documentazione API",
    "version": "1.0"
  },
  "paths": {
    "/api/calendar/{user}": {
      "get": {
        "description": "Ritorna tutti i calendari di un utentente dato l'userId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "UserId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "examples": {
              "Utente1": {
                "summary": "Utente1",
                "value": "User1"
              },
              "Utente2": {
                "summary": "Utente2",
                "value": "User2"
              },
              "Utente3": {
                "summary": "Utente3",
                "value": "User3"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "E' stato trovato almeno un calendario e verra restituito sotto forma di lista",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "calendari": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "nome": {
                            "type": "string"
                          },
                          "fusoOrario": {
                            "type": "object",
                            "properties": {
                              "GMTOffset": {
                                "type": "integer",
                                "minimum": -12,
                                "maximum": 12
                              },
                              "localita": {
                                "type": "string"
                              }
                            }
                          },
                          "colore": {
                            "type": "string",
                            "pattern": "/^#([0-9a-f]{3}){1,2}$/i"
                          },
                          "partecipanti": {
                            "type": "array",
                            "items": {
                              "type": "string",
                              "format": "userId"
                            }
                          },
                          "principale": {
                            "type": "boolean"
                          },
                          "impostazioniPredefiniteEventi": {
                            "type": "object",
                            "properties": {
                              "titolo": {
                                "type": "string"
                              },
                              "descrizione": {
                                "type": "string"
                              },
                              "durata": {
                                "type": "integer",
                                "minimum": 1
                              },
                              "tempAnticNotifica": {
                                "type": "integer",
                                "minimum": 0
                              },
                              "luogo": {
                                "type": "object",
                                "properties": {
                                  "latitudine": {
                                    "type": "string"
                                  },
                                  "longitudine": {
                                    "type": "string"
                                  }
                                }
                              },
                              "priorita": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 10
                              },
                              "difficolta": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 10
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "Compleanno": {
                    "summary": "Esempio di un calendario per compleanni",
                    "value": {
                      "nome": "Compleanni",
                      "fusoOrario": {
                        "GMTOffset": 0,
                        "localita": "London"
                      },
                      "colore": "#FFAAFF",
                      "partecipanti": [
                        "User1",
                        "User2",
                        "User3"
                      ],
                      "principale": false,
                      "impostazioniPredefiniteEventi": {
                        "titolo": "Compleanno",
                        "descrizione": "Compleanno di",
                        "durata": 1440,
                        "tempAnticNotifica": 1440,
                        "luogo": {
                          "latitudine": "12N",
                          "longitudine": "13E"
                        },
                        "priorita": 7,
                        "difficolta": 1
                      }
                    }
                  },
                  "Lavoro": {
                    "summary": "Esempio di un calendario di lavoro",
                    "value": {
                      "nome": "Lavoro",
                      "fusoOrario": {
                        "GMTOffset": -5,
                        "localita": "NewYork"
                      },
                      "colore": "#FF0000",
                      "partecipanti": [
                        "User1",
                        "User3"
                      ],
                      "principale": false,
                      "impostazioniPredefiniteEventi": {
                        "titolo": "Meeting",
                        "descrizione": "Aggiornamento su ",
                        "durata": 30,
                        "tempAnticNotifica": 10,
                        "luogo": {
                          "latitudine": "12N",
                          "longitudine": "1E"
                        },
                        "priorita": 5,
                        "difficolta": 6
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Manca il parametro userId, verra restituito \"Parameter missing\"",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Parameter missing"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Se non e' stato trovato neanche un account con l'userId dato, verra restituito \"There is no user with that userId\" <br>Se e' stato trovato piu di un account con lo stesso userId, verra restituito \"There are too many users with that userId\" <br>Se non sono stati trovati calendari per l'userId dato, verra restituito \"There are no calendars with that userId\"",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "Errore1": {
                    "summary": "Neanche un utente con l'userId specificato",
                    "value": {
                      "error": "There is no user with that userId"
                    }
                  },
                  "Errore2": {
                    "summary": "Piu utenti con lo stesso userId",
                    "value": {
                      "error": "There are too many users with that userId"
                    }
                  },
                  "Errore3": {
                    "summary": "Utente senza calendari",
                    "value": {
                      "error": "There are no calendars with that userId"
                    }
                  }
                }
              }
            }
          },
          "501": {
            "description": "Errore generico, verra restituito \"Generic error\"",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Generic error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}